UI improvements: icons, floating boxes, panel resizing and consuming input
major:
- Input->Consumed flag to have input be consumed by only one UI element
  at the time.
- Reworked and improved panel resizing, dragging info is provided in
  Input (StartX, StartY)
- Add floating ui boxes
- Added second font for icons

misc:
- (Re)Enabled sign conversion warnings
- Option to keep old DLL around, disabled to play nice with raddbg.
  Keeping it is no longer needed since we are using readonly pages.
- fix ui violation undead code bug
- Removed titlebar
- fix cling metaprogram and run the metaprogram on build
- set clipboard with boolean to avoid second copy

+++ source/base/base_strings.c
@@ -132,7 +132,7 @@ Str8Fmt(char *Format, ...)
     va_list Args;
     va_start(Args, Format);
     
-    Result.Size = vsprintf((char *)Result.Data, Format, Args);
+    Result.Size = (u64)vsprintf((char *)Result.Data, Format, Args);
     
     return Result;
 }
